[nvfp4_training] Parametrize the NVFP4 tests over the kernel backend
Per review feedback ("make the tests parametrize over the kernel selector
instead of copy paste"), collapse the eight per-backend test files into four,
each parametrized over the kernel via a _KERNELS list of pytest.param entries
carrying per-backend skip marks -- mirroring the bench merge. This removes the
duplicated reference/helper functions (_rht_reference, _dequantize,
_quantize_row_col, etc.) that were copy-pasted between the _triton and _cutedsl
members of each pair.

  test_hadamard_amax_{triton,cutedsl}.py            -> test_hadamard_amax.py
  test_hadamard_quantize_row_col_{triton,cutedsl}.py -> test_hadamard_quantize_row_col.py
  test_quantize_2d_{triton,cutedsl}.py              -> test_quantize_2d.py
  test_nvfp4_linear{,_cutedsl}.py                   -> test_nvfp4_linear.py

Tests that are the same computation modulo the op are parametrized over the
backend; genuinely backend-specific tests (e.g. the cutedsl shape/dtype
validation, the triton stochastic-rounding statistics) keep a single-backend
skip; and cross-backend agreement tests (cutedsl-vs-triton) stay as one test
requiring both. Assertions are preserved per backend -- where the two differ
(triton matches a reference bitwise, cutedsl up to a tolerance / FP8 rounding
ties) the test branches on the kernel rather than relaxing the stricter side.

Where the backends accept different shapes, the sweep is the union and a
_skip_if_unsupported_shape guard skips what a backend can't run, so triton keeps
its sub-256 M coverage without a duplicated test body.
