+++ test/prototype/moe_training/nvfp4_training/test_quantize_2d.py
@@ -1,37 +1,78 @@
-"""Tests for triton_weight_quantize_2d (SM100+ kernel).
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD 3-Clause license found in the
+# LICENSE file in the root directory of this source tree.
 
-test_triton_weight_quantize_2d_scales_vs_reference:
-  FP8 scale factors must match the PyTorch reference bitwise for both
-  non-swizzled (M, N//16) and swizzled (M//128, N//64, 32, 16) layouts.
-
-test_triton_weight_quantize_2d_sqnr:
-  Dequantized output must reconstruct A with SQNR >= 20 dB.
-"""
 
 import pytest
 import torch
 from torch.utils._triton import has_triton
 
 from torchao.float8.float8_utils import compute_error
+from torchao.prototype.moe_training.nvfp4_training.hadamard_cutedsl_utils import (
+    cutedsl_nvfp4_kernels_available,
+)
+from torchao.prototype.moe_training.nvfp4_training.hadamard_utils import (
+    prepare_for_cuda_graph,
+)
+from torchao.prototype.moe_training.nvfp4_training.quantize_2d_cutedsl import (
+    cutedsl_weight_quantize_2d,
+)
 from torchao.prototype.mx_formats.nvfp4_tensor import (
     NVFP4Tensor,
     per_tensor_amax_to_scale,
 )
 from torchao.utils import is_sm_at_least_100, torch_version_at_least
 
-if has_triton() and is_sm_at_least_100() and torch_version_at_least("2.10.0"):
-    from torchao.prototype.moe_training.nvfp4_training.quantize_2d_triton import (
-        triton_weight_quantize_2d,
-    )
-from torchao.prototype.moe_training.nvfp4_training.hadamard_utils import (
-    prepare_for_cuda_graph,
+_skip_no_triton = pytest.mark.skipif(
+    not (has_triton() and is_sm_at_least_100() and torch_version_at_least("2.10.0")),
+    reason="requires triton, SM100+, and PyTorch 2.10+ (torch.compile)",
+)
+_skip_no_cutedsl = pytest.mark.skipif(
+    not cutedsl_nvfp4_kernels_available(),
+    reason="requires SM100 (Blackwell) + CuteDSL runtime (cuda-python, nvidia-cutlass-dsl)",
 )
 
-# BLOCK_M minimum is 128; N must be a multiple of BLOCK_N=256.
+_KERNELS = [
+    pytest.param("triton", marks=_skip_no_triton, id="triton"),
+    pytest.param("cutedsl", marks=_skip_no_cutedsl, id="cutedsl"),
+]
+
+# The CuteDSL kernel requires out_features (dim 0) % 256 == 0 and in_features (dim 1)
+# % 128 == 0; the triton kernel's BLOCK_M minimum is 128 and its swizzled scales require
+# both dims % 128. The sweep is the union, with _skip_if_unsupported_shape dropping
+# M=128 for cutedsl so triton keeps its BLOCK_M-minimum coverage.
 _M_VALUES = [128, 256, 512, 1024]
-_N_VALUES = [256, 512, 1024, 2048]
+_N_VALUES = [128, 256, 512, 1024, 2048]
+
+
+def _skip_if_unsupported_shape(kernel: str, M: int, N: int) -> None:
+    """Skip shapes the selected backend cannot handle."""
+    if kernel == "cutedsl" and M % 256 != 0:
+        pytest.skip("cutedsl weight quantize requires out_features % 256 == 0")
+
+
+# Minimum reconstruction SQNR (dB) per backend; both land around 19 dB on the grid above.
+_MIN_SQNR_DB = {"triton": 15.0, "cutedsl": 18.0}
+
 _FP8_E4M3_EPS = torch.finfo(torch.float8_e4m3fn).tiny
 _NEAR_ZERO = 1.0e-10
+
+
+def _weight_quantize_2d(kernel, W, amax):
+    """Dispatch to a backend's 2D weight-quantize kernel.
+
+    Returns ``(codes, scales, t_codes, t_scales)``: the rowwise NVFP4(W) codes + swizzled
+    FP8 scale factors, then the colwise NVFP4(W.T) codes + swizzled scale factors.
+    """
+    if kernel == "triton":
+        from torchao.prototype.moe_training.nvfp4_training.quantize_2d_triton import (
+            triton_weight_quantize_2d,
+        )
+
+        return triton_weight_quantize_2d(W, amax)
+    return cutedsl_weight_quantize_2d(W, amax)
 
 
 # ---------------------------------------------------------------------------
@@ -156,116 +197,151 @@ def _assert_near_zero_values_do_not_saturate(
     )
 
 
+def _assert_scales_match_up_to_rounding_ties(
+    scales: torch.Tensor, reference: torch.Tensor, what: str
+) -> None:
+    """Allow the small fraction of blocks whose FP8 scale rounds the tie the other way.
+
+    Every mismatch must be within one E4M3 step (ratio in [1/1.15, 1.15]), so a real recipe
+    divergence (e.g. 1D instead of 2D block scaling) still fails.
+    """
+    got = scales.flatten().float()
+    ref = reference.flatten().float()
+    mism = got != ref
+    frac = mism.float().mean().item()
+    assert frac < 0.02, f"{100 * frac:.2f}% of {what} differ (>2%)"
+    if mism.any():
+        ratio = got[mism] / ref[mism]
+        assert (ratio <= 1.15).all() and (ratio >= 1 / 1.15).all(), (
+            f"{what} mismatches exceed one E4M3 step (not a rounding tie)"
+        )
+
+
+def _assert_scales_vs_reference(
+    kernel: str, scales: torch.Tensor, reference: torch.Tensor, what: str
+) -> None:
+    """triton reproduces the PyTorch oracle bitwise; CuteDSL up to FP8 rounding ties."""
+    if kernel == "triton":
+        torch.testing.assert_close(scales, reference, atol=0, rtol=0)
+    else:
+        _assert_scales_match_up_to_rounding_ties(scales, reference, what)
+
+
 # ---------------------------------------------------------------------------
 # Tests — scale factors
 # ---------------------------------------------------------------------------
 
 
-@pytest.mark.skipif(not has_triton(), reason="unsupported without triton")
-@pytest.mark.skipif(not is_sm_at_least_100(), reason="Requires SM100+")
-@pytest.mark.skipif(
-    not torch_version_at_least("2.10.0"), reason="torch.compile requires PyTorch 2.10+"
-)
+@pytest.mark.parametrize("kernel", _KERNELS)
 @pytest.mark.parametrize("N", _N_VALUES, ids=lambda n: f"N{n}")
 @pytest.mark.parametrize("M", _M_VALUES, ids=lambda m: f"M{m}")
 @torch.no_grad()
-def test_triton_weight_quantize_2d_scales_vs_reference(M, N):
-    """FP8 scale factors must match the PyTorch reference bitwise."""
-    if M % 128 != 0 or N % 128 != 0:
-        pytest.skip("swizzled scales require M % 128 == 0 and N % 128 == 0")
-
+def test_weight_quantize_2d_scales_vs_reference(kernel, M, N):
+    """Swizzled FP8 scale factors must match the PyTorch 16x16 reference."""
+    _skip_if_unsupported_shape(kernel, M, N)
     torch.manual_seed(42)
     A = torch.randn(M, N, dtype=torch.bfloat16, device="cuda")
 
     ref_scales_expanded = _weight_quantize_2d_reference_scales(A)  # (M, N//16)
 
-    _, tri_scales, _, tri_t_scales = triton_weight_quantize_2d(A, A.float().abs().max())
+    _, scales, _, t_scales = _weight_quantize_2d(kernel, A, A.float().abs().max())
 
     ref_scales = _swizzle_py(ref_scales_expanded, M, N)
-    torch.testing.assert_close(tri_scales, ref_scales, atol=0, rtol=0)
+    _assert_scales_vs_reference(kernel, scales, ref_scales, "rowwise SF")
 
     ref_t_scales_expanded = _weight_quantize_2d_reference_scales(
         A.T.contiguous()
     )  # (N, M//16)
     ref_t_scales = _swizzle_py(ref_t_scales_expanded, N, M)  # (N//128, M//64, 32, 16)
-    torch.testing.assert_close(tri_t_scales, ref_t_scales, atol=0, rtol=0)
+    _assert_scales_vs_reference(kernel, t_scales, ref_t_scales, "colwise SF")
 
 
-# ---------------------------------------------------------------------------
-# Tests — quantization quality (SQNR)
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.skipif(not has_triton(), reason="unsupported without triton")
-@pytest.mark.skipif(not is_sm_at_least_100(), reason="Requires SM100+")
-@pytest.mark.skipif(
-    not torch_version_at_least("2.10.0"), reason="torch.compile requires PyTorch 2.10+"
-)
+@_skip_no_triton
+@_skip_no_cutedsl
 @pytest.mark.parametrize("N", _N_VALUES, ids=lambda n: f"N{n}")
 @pytest.mark.parametrize("M", _M_VALUES, ids=lambda m: f"M{m}")
 @torch.no_grad()
-def test_triton_weight_quantize_2d_sqnr(M, N):
-    """Dequantized output must reconstruct A with SQNR >= 20 dB."""
-    if M % 128 != 0 or N % 128 != 0:
-        pytest.skip("swizzled scales require M % 128 == 0 and N % 128 == 0")
+def test_cutedsl_weight_quantize_2d_matches_triton(M, N):
+    """Both backends emit the same 2D 16x16 scale factors, modulo FP8 rounding ties."""
+    _skip_if_unsupported_shape("cutedsl", M, N)
+    torch.manual_seed(3)
+    W = torch.randn(M, N, dtype=torch.bfloat16, device="cuda")
+    amax = W.float().abs().max()
+    _, c_rsf, _, c_csf = _weight_quantize_2d("cutedsl", W, amax)
+    _, t_rsf, _, t_csf = _weight_quantize_2d("triton", W, amax)
+    _assert_scales_match_up_to_rounding_ties(c_rsf, t_rsf, "rowwise SF vs Triton 16x16")
+    _assert_scales_match_up_to_rounding_ties(c_csf, t_csf, "colwise SF vs Triton 16x16")
 
+
+# ---------------------------------------------------------------------------
+# Tests — quantization quality (SQNR) and output contract
+# ---------------------------------------------------------------------------
+
+
+@pytest.mark.parametrize("kernel", _KERNELS)
+@pytest.mark.parametrize("N", _N_VALUES, ids=lambda n: f"N{n}")
+@pytest.mark.parametrize("M", _M_VALUES, ids=lambda m: f"M{m}")
+@torch.no_grad()
+def test_weight_quantize_2d_sqnr(kernel, M, N):
+    """Rowwise dequant must reconstruct A and colwise dequant A.T."""
+    _skip_if_unsupported_shape(kernel, M, N)
     torch.manual_seed(42)
     A = torch.randn(M, N, dtype=torch.bfloat16, device="cuda")
 
     global_amax = A.float().abs().max()
-    tri_codes, tri_scales, tri_t_codes, tri_t_scales = triton_weight_quantize_2d(
-        A, global_amax
-    )
+    codes, scales, t_codes, t_scales = _weight_quantize_2d(kernel, A, global_amax)
 
     # NVFP4Tensor interprets (M, N//16) scales as rowwise block_size=16 scales,
     # which matches the 2D expanded layout since every 16 rows share the same scale.
-    dequant = (
-        NVFP4Tensor(
-            tri_codes,
-            tri_scales,
-            16,
-            torch.bfloat16,
-            per_tensor_scale=per_tensor_amax_to_scale(global_amax),
-            is_swizzled_scales=True,
-        )
-        .dequantize()
-        .float()
+    min_sqnr = _MIN_SQNR_DB[kernel]
+
+    sqnr = compute_error(A.float(), _dequantize(codes, scales, global_amax))
+    assert sqnr >= min_sqnr, (
+        f"Rowwise SQNR {sqnr:.2f} dB < {min_sqnr} dB for M={M} N={N}"
     )
 
-    sqnr = compute_error(A.float(), dequant)
-    assert sqnr >= 15.0, f"Rowwise SQNR {sqnr:.2f} dB < 15.0 dB for M={M} N={N}"
-
-    dequant_t = (
-        NVFP4Tensor(
-            tri_t_codes,
-            tri_t_scales,
-            16,
-            torch.bfloat16,
-            per_tensor_scale=per_tensor_amax_to_scale(global_amax),
-            is_swizzled_scales=True,
-        )
-        .dequantize()
-        .float()
+    sqnr_t = compute_error(A.T.float(), _dequantize(t_codes, t_scales, global_amax))
+    assert sqnr_t >= min_sqnr, (
+        f"Colwise SQNR {sqnr_t:.2f} dB < {min_sqnr} dB for M={M} N={N}"
     )
 
-    sqnr_t = compute_error(A.T.float(), dequant_t)
-    assert sqnr_t >= 15.0, f"Colwise SQNR {sqnr_t:.2f} dB < 15.0 dB for M={M} N={N}"
+
+@pytest.mark.parametrize("kernel", _KERNELS)
+@pytest.mark.parametrize("N", _N_VALUES, ids=lambda n: f"N{n}")
+@pytest.mark.parametrize("M", _M_VALUES, ids=lambda m: f"M{m}")
+@torch.no_grad()
+def test_weight_quantize_2d_layout(kernel, M, N):
+    """Output 4-tuple shapes + dtypes, the contract both backends share."""
+    _skip_if_unsupported_shape(kernel, M, N)
+    torch.manual_seed(1)
+    W = torch.randn(M, N, dtype=torch.bfloat16, device="cuda")
+    amax = W.float().abs().max()
+    w_fp4, w_sf, wt_fp4, wt_sf = _weight_quantize_2d(kernel, W, amax)
+
+    assert w_fp4.shape == (M, N // 2) and w_fp4.dtype == torch.uint8
+    assert wt_fp4.shape == (N, M // 2) and wt_fp4.dtype == torch.uint8
+    assert (
+        w_sf.shape == (M // 128, N // 64, 32, 16) and w_sf.dtype == torch.float8_e4m3fn
+    )
+    assert (
+        wt_sf.shape == (N // 128, M // 64, 32, 16)
+        and wt_sf.dtype == torch.float8_e4m3fn
+    )
 
 
-@pytest.mark.skipif(not has_triton(), reason="unsupported without triton")
-@pytest.mark.skipif(not is_sm_at_least_100(), reason="Requires SM100+")
-@pytest.mark.skipif(
-    not torch_version_at_least("2.10.0"), reason="torch.compile requires PyTorch 2.10+"
-)
+# ---------------------------------------------------------------------------
+# Tests — degenerate inputs
+# ---------------------------------------------------------------------------
+
+
+@pytest.mark.parametrize("kernel", _KERNELS)
 @pytest.mark.parametrize(
     "input_kind",
     [pytest.param("zeros", id="zeros"), pytest.param("near_zero", id="near_zero")],
 )
 @torch.no_grad()
-def test_triton_weight_quantize_2d_zero_and_near_zero_no_nan_or_saturation(
-    input_kind,
-):
-    M, N = 128, 256
+def test_weight_quantize_2d_zero_and_near_zero_no_nan_or_saturation(kernel, input_kind):
+    M, N = 256, 256
 
     if input_kind == "zeros":
         A = torch.zeros(M, N, dtype=torch.bfloat16, device="cuda")
@@ -274,9 +350,11 @@ def test_triton_weight_quantize_2d_zero_and_near_zero_no_nan_or_saturation(
         A[0, 0] = 1.0
 
     global_amax = A.float().abs().max()
-    row_codes, row_sf, col_codes, col_sf = triton_weight_quantize_2d(A, global_amax)
+    row_codes, row_sf, col_codes, col_sf = _weight_quantize_2d(kernel, A, global_amax)
 
     if input_kind == "zeros":
+        # Zero input packs to zero codes, every block scale clamps to E4M3 eps (not 0),
+        # and both layouts dequantize back to zero.
         _assert_zero_quantized(
             row_codes, row_sf, _dequantize(row_codes, row_sf, global_amax)
         )
@@ -308,18 +386,30 @@ def test_triton_weight_quantize_2d_zero_and_near_zero_no_nan_or_saturation(
     _assert_near_zero_values_do_not_saturate(col_codes, col_near_zero_mask)
 
 
-@pytest.mark.skipif(not has_triton(), reason="unsupported without triton")
-@pytest.mark.skipif(not is_sm_at_least_100(), reason="Requires SM100+")
-@pytest.mark.skipif(
-    not torch_version_at_least("2.10.0"), reason="torch.compile requires PyTorch 2.10+"
-)
+# ---------------------------------------------------------------------------
+# Tests — backend-specific
+# ---------------------------------------------------------------------------
+
+
+@_skip_no_cutedsl
+@torch.no_grad()
+def test_cutedsl_weight_quantize_2d_requires_out_features_256():
+    """out_features (dim 0) must be divisible by 256 (stricter than the Triton kernel's 128)."""
+    W = torch.randn(384, 256, dtype=torch.bfloat16, device="cuda")  # 384 % 256 == 128
+    amax = W.float().abs().max()
+    with pytest.raises(ValueError, match="out_features"):
+        cutedsl_weight_quantize_2d(W, amax)
+
+
+@_skip_no_triton
 @pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA not available")
 def test_triton_weight_quantize_2d_cuda_graph_compile():
-    """triton_weight_quantize_2d_colwise under reduce-overhead CUDA graphs.
+    """triton_weight_quantize_2d under reduce-overhead CUDA graphs.
 
     Weight quantization is deterministic (no SR), so consecutive calls should produce
     identical outputs. Primarily verifies fullgraph=True compilation succeeds via the
-    registered custom_op + register_fake.
+    registered custom_op + register_fake. Also the only M % 256 != 0 coverage, a shape
+    the CuteDSL kernel rejects.
     """
     shape = (128, 256)
     W = torch.randn(*shape, dtype=torch.bfloat16, device="cuda")
@@ -329,7 +419,7 @@ def test_triton_weight_quantize_2d_cuda_graph_compile():
 
     def run(w):
         amax = w.float().abs().max()
-        codes, sf, t_codes, t_sf = triton_weight_quantize_2d(w, amax)
+        codes, sf, t_codes, t_sf = _weight_quantize_2d("triton", w, amax)
         return codes.clone()
 
     compiled = torch.compile(run, mode="reduce-overhead", fullgraph=True)
